Add Reducer and Actions
