+++ src/index.css
@@ -5,8 +5,8 @@ body {
 }
 
 .container {
-  margin: 0 2.5rem 4.063rem 0;
-  padding: 0 0 49.625rem;
+  margin: 0 10px 10px 0;
+  padding: 0 0 40px;
   border-radius: 4px;
   box-shadow: 0 0 60px 0 rgba(0, 0, 0, 0.05);
   background-color: #fafafa;
