+++ src/index.js
@@ -1,14 +1,19 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from "react-router-dom";
+import { Provider } from "react-redux";
+import configureStore from "./redux/configureStore";
 import './index.css';
 import App from './components/Index';
 
 ReactDOM.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <Provider store={configureStore}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </Provider>
+
 
   </React.StrictMode>,
   document.getElementById('root')
