+++ src/pages/Categories.js
@@ -3,19 +3,8 @@ import React from 'react';
 
 const Categories = () => (
   <>
-    <div className="container">
-      <h2 className="title"> Welcome to our page!</h2>
-      <div>
-        <p className="content">
-          Lorem Ipsum is simply dummy text of the printing and typesetting industry.
-          Lorem Ipsum has been the industrys standard dummy text ever since the 1500s,
-          when an unknown printer took a galley of type and scrambled it to make a type specimen.
-          It has survived not only five centuries, but also the leap into electronic typesetting,
-          remaining essentially unchanged. It was popularised in the 1960s with the release of
-          Letraset sheets containing Lorem Ipsum passages, and more recently with desktop publishing
-          software like Aldus PageMaker including versions of Lorem Ipsum.
-        </p>
-      </div>
+    <div>
+      <h2> Under construction</h2>
     </div>
   </>
 );
