+++ src/redux/books/actions.js
@@ -1,0 +1,3 @@
+export const addBook = payload => ({
+    type: "ADDBOOK"
+});
