+++ src/redux/books/reducers.js
@@ -1,0 +1,6 @@
+export const booksReducer = (state = null, action) => {
+    switch (action.type) {
+        default:
+            return state;
+    }
+}
