+++ src/redux/categories/reducers.js
@@ -1,0 +1,6 @@
+export const categoriesReducer = (state = null, action) => {
+    switch (action.type) {
+        default:
+            return state;
+    }
+}
